requirements.txt
torch>=2.1
transformers>=4.40
peft>=0.8
pandas>=2.0
numpy>=1.24
tqdm
scikit-learn
pyyaml
matplotlib
